pyproject.toml
[tool.poetry]
name = "isi-minimal"
version = "0.1.0"
description = "Minimal ISI Macroscope Control System"
authors = ["ISI Lab"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.24.0"
scipy = "^1.10.0"
h5py = "^3.8.0"
opencv-python = "^4.7.0"
pillow = "^10.0.0"
websockets = "^11.0.0"
pyzmq = "^27.1.0"
psutil = "^7.1.0"
pyglet = "^2.0.0"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"